requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torch-audiomentations>=0.11.0
pytorch-lightning>=2.0.0
git+https://github.com/joonaskalda/pyannote-audio.git@dgmse
pyannote.database>=5.0.0
